pyproject.toml
[project]
name = "visual-engine-mcp"
version = "0.1.0"
description = "MCP server for AI-powered image analysis using Google Gemini vision"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "google-genai>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "httpx>=0.24.0",
]

[project.scripts]
visual-engine = "visual_engine.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["visual_engine"]
